crypto/boring: Refactor boringcrypto codebase
This patch refactors the boringcrypto implementation in an attempt to
make it more general and abstract away the actual external crypto
backend being used.

This is in effort to enable other external crypto libraries such as
OpenSSL to be used instead.
