+++ src/cmd/link/internal/ld/lib.go
@@ -1012,6 +1012,7 @@ var hostobj []Hostobj
 // Others trigger external mode.
 var internalpkg = []string{
 	"crypto/internal/boring",
+	"crypto/internal/boring/boringcrypto",
 	"crypto/x509",
 	"net",
 	"os/user",
