+++ src/crypto/internal/boring/rand.go
@@ -9,19 +9,6 @@
 
 package boring
 
-// #include "goboringcrypto.h"
-import "C"
-import "unsafe"
+import "crypto/internal/boring/boringcrypto"
 
-type randReader int
-
-func (randReader) Read(b []byte) (int, error) {
-	// Note: RAND_bytes should never fail; the return value exists only for historical reasons.
-	// We check it even so.
-	if len(b) > 0 && C._goboringcrypto_RAND_bytes((*C.uint8_t)(unsafe.Pointer(&b[0])), C.size_t(len(b))) == 0 {
-		return 0, fail("RAND_bytes")
-	}
-	return len(b), nil
-}
-
-const RandReader = randReader(0)
+const RandReader = boringcrypto.RandReader(0)
